♻️ refactor: modify component path #7
